TemporalGPs.jl

ChainRulesCore.ProjectTo

TemporalGPs.AbstractLGC

TemporalGPs.BottleneckLGC

TemporalGPs.DTCSeparable

TemporalGPs.ElementOfLGSSM

TemporalGPs.FiniteLTISDE

TemporalGPs.GaussMarkovModel

TemporalGPs.Gaussian

TemporalGPs.LGSSM

TemporalGPs.LTISDE

TemporalGPs.LargeOutputLGC

TemporalGPs.RectilinearGrid

TemporalGPs.RegularInTime

TemporalGPs.RegularSpacing

TemporalGPs.ScalarOutputLGC

TemporalGPs.Separable

TemporalGPs.SmallOutputLGC

TemporalGPs.SpaceTimeGrid

AbstractGPs.dtc

AbstractGPs.elbo

AbstractGPs.marginals

TemporalGPs.approx_posterior_marginals

TemporalGPs.approx_posterior_marginals

TemporalGPs.conditional_rand

TemporalGPs.destructure

TemporalGPs.dtcify

TemporalGPs.get_times

TemporalGPs.inputs_to_time_form

TemporalGPs.marginals_diag

TemporalGPs.merge_inputs

TemporalGPs.noise_var_to_time_form

TemporalGPs.observations_to_time_form

TemporalGPs.predict

TemporalGPs.predict_marginals

TemporalGPs.scan_emit

TemporalGPs.sort_in_time

TemporalGPs.to_sde

TemporalGPs.to_sde

TemporalGPs.zygote_friendly_map

TemporalGPs.ε_randn

ChainRulesCore.ProjectTo — Method

ProjectTo(::Fill) -> ProjectTo{Fill}
ProjectTo(::Ones) -> ProjectTo{NoTangent}

Most FillArrays arrays store one number, and so their gradients under automatic differentiation represent the variation of this one number.

The exception is those like Ones and Zeros whose type fixes their value, which have no graidient.

TemporalGPs.AbstractLGC — Type

AbstractLGC

Represents a Gaussian conditional distribution:

y | x ∼ Gaussian(A * x + a, Q)

Note that this can be used in two contexts:

Transition: y is the next state, x is the current state.

Emission: y is the observation, x is the state.

Subtypes have discretion over how to implement the interface for this type. In particular A, a, and Q may not be represented explicitly so that structure can be exploited to accelerate computations.

Interface:

==

eltype

dim_out

dim_in

predict

predict_marginals

conditional_rand

ε_randn

posterior_and_lml

TemporalGPs.BottleneckLGC — Type

BottleneckLGC

A composition of an affine map that projects onto a low-dimensional subspace and a LargeOutputLGC. This structure is exploited by only ever computing Cholesky factorisations in the space the affine map maps to, rather than the input or output space.

Letting, H and h parametrise the affine map, and f the "fan-out" LargeOutputLGC, the conditional distribution that this model parametrises is

y | x ~ Gaussian(f.A * (H * x + h) + f.a, f.Q)

Note that this type does not enforce that size(H, 1) < size(H, 2), nor that dim_out(f) > dim_in(f), it's just not a particularly good idea to use this type unless your model satisfies these properties.

TemporalGPs.DTCSeparable — Type

DTCSeparable{Tz<:AbstractVector, Tk<:SeparableKernel} <: Kernel

Specifies a low-rank approximation to a kernel k through pseudo-inputs z. z are the locations of the pseudo-inputs in space, since they are automatically replicated (implicitly) at each point in time.

TemporalGPs.ElementOfLGSSM — Type

ElementOfLGSSM

Represents an element of LGSSM with a given ordering.

TemporalGPs.FiniteLTISDE — Type

const FiniteLTISDE = FiniteGP{<:LTISDE}

A FiniteLTISDE is just a regular FiniteGP that happens to contain an LTISDE, as opposed to any other AbstractGP, useful for dispatching.

TemporalGPs.GaussMarkovModel — Type

GaussMarkovModel

Specifies the following Gauss-Markov model.

x[0] ∼ x0
x[t] = A[t] * x[t-1] + a[t] + ε[t], ε[t] ∼ N(0, Q)

TemporalGPs.Gaussian — Type

Gaussian{Tm, TP}

A (multivariate) Gaussian with mean vector m and variance matrix P. This doesn't currently conform to Distributions.jl standards. Work to make this happen would be welcomed.

It was necessary to implement this a year or so ago for AD-related reasons. It's quite possible that in the intervening period of time things have improved and this type is no longer necessary in addition to the MvNormal type in Distributions. I've not had the time to remove it though.

TemporalGPs.LGSSM — Type

LGSSM{Ttransitions<:GaussMarkovModel, Temissions<:StructArray} <: AbstractLGSSM

A Linear-Gaussian State-Space model. Represented in terms of a Gauss-Markov model transitions and collection of emission dynamics emissions.

TemporalGPs.LTISDE — Type

LTISDE (Linear Time-Invariant Stochastic Differential Equation)

A lightweight wrapper around a GP f that tells this package to handle inference in f. Can be constructed via the to_sde function.

TemporalGPs.LargeOutputLGC — Type

LargeOutputLGC{
    TA<:AbstractMatrix, Ta<:AbstractVector, TQ<:AbstractMatrix,
} <: AbstractLGC

A SmallOutputLGC (LGC) specialised for models in which the dimension of the outputs are greater than that of the inputs. These specialisations both improve numerical stability and performance (time and memory), so it's worth using if your model lives in this regime.

TemporalGPs.RectilinearGrid — Type

RectilinearGrid{Tl, Tr} <: AbstractVector{Tuple{Tl, Tr}}

A RectilinearGrid is parametrised by AbstractVectors of points xl and xr, whose element types are Tl and Tr respectively, comprising length(xl) * length(xr) elements. Linear indexing is the same as product(eachindex(xl), eachindex(xr)) - xl iterates more quickly than xr.

TemporalGPs.RegularInTime — Type

RegularInTime{
    Tt, Tv, Tts<:AbstractVector{Tt}, Tvs<:AbstractVector{<:AbstractVector{Tv}},
} <: AbstractVector{Tuple{Tt, Tv}}

Represents data that has multiple observations at each of a given collection of time slices.

TemporalGPs.RegularSpacing — Type

RegularSpacing{T<:Real} <: AbstractVector{T}

RegularSpacing(t0, Δt, N) represents the same thing as range(t0; step=Δt, length=N), but has a different implementation that makes it possible to differentiate through with the current version of Zygote. This data structure will be entirely removed once it's possible to work with StepRangeLens in Zygote.

Relevant issue: https://github.com/FluxML/Zygote.jl/issues/550

TemporalGPs.ScalarOutputLGC — Type

ScalarOutputLGC

An AbstractLGC that operates on a vector-valued input space and a scalar-valued output space. Similar to SmallOutputLGC when its dim_out is 1 but, for example, conditional_rand returns a Real rather than an AbstractVector of length 1.

TemporalGPs.Separable — Type

Separable{Tl<:Kernel, Tr<:Kernel} <: Kernel

The kernel k given by

k((xl, xr), (yl, yr)) = k.l(xl, yl) * k.r(xr, yr)

TemporalGPs.SmallOutputLGC — Type

SmallOutputLGC{
    TA<:AbstractMatrix, Ta<:AbstractVector, TQ<:AbstractMatrix,
} <: AbstractLGC

a.k.a. LGC. An AbstractLGC designed for problems in which A is a matrix, and size(A, 1) < size(A, 2). It should still work (roughly) for problems in which size(A, 1) > size(A, 2), but one should expect worse accuracy and performance than a LargeOutputLGC in such circumstances.

TemporalGPs.SpaceTimeGrid — Type

SpaceTimeGrid{Tr, Tt<:Real}

A SpaceTimeGrid is a RectilinearGrid in which the left vector corresponds to space, and the right time. The left eltype is arbitrary, but the right must be Real.

AbstractGPs.dtc — Method

dtc(fx::FiniteLTISDE, y::AbstractVector{<:Real}, z_r::AbstractVector)

Compute the DTC (Deterministic Training Conditional) in state-space form [insert reference].

fx and y are the same as would be provided to logpdf, and z_r is a specification of the spatial location of the pseudo-points at each point in time.

Note that this API is slightly different from AbstractGPS.jl's API, in which z_r is replaced by a FiniteGP.

WARNING: this API is unstable, and subject to change in future versions of TemporalGPs. It was thrown together quickly in pursuit of a conference deadline, and has yet to receive the attention it deserves.

AbstractGPs.elbo — Method

elbo(fx::FiniteLTISDE, y::AbstractVector{<:Real}, z_r::AbstractVector)

Compute the ELBO (Evidence Lower BOund) in state-space form [insert reference].

AbstractGPs.marginals — Method

marginals(model::LGSSM)

Compute the complete marginals at each point in time. These are returned as a Vector of length length(model), each element of which is a dense Gaussian.

TemporalGPs.approx_posterior_marginals — Method

approx_posterior_marginals(
    ::typeof(dtc),
    fx::FiniteLTISDE,
    y::AbstractVector,
    z_r::AbstractVector,
    x_r::AbstractVector,
    t::Int,
)

Same as other method, but only returns the predictions at index t in fx.

As with the other method of this function, it's a bit of a hack. It's correct of course, but needs to be tidied up at some point.

TemporalGPs.approx_posterior_marginals — Method

approx_posterior_marginals(
    ::typeof(dtc),
    fx::FiniteLTISDE,
    y::AbstractVector,
    z_r::AbstractVector,
    x_r::AbstractVector,
)

Compute the DTC (Deterministic Training Conditional) approximation to the posterior marginals at the times provided by fx, but at the new spatial locations x_r, given observations y and spatial pseudo-input locations z_r.

WARNING: this API is unstable, and subject to change in future versions of TemporalGPs. It was thrown together quickly in pursuit of a conference deadline, and has yet to receive the attention it deserves.

TemporalGPs.conditional_rand — Method

conditional_rand(rng::AbstractRNG, f::AbstractLGC, x::AbstractVector)
conditional_rand(ε::AbstractVector, f::AbstractLGC, x::AbstractVector)

Sample from the conditional distribution y | x. ε is the randomness needed to generate this sample. If rng is provided, it will be used to construct ε via ε_randn.

If implementing a new AbstractLGC, implement the ε method as it avoids randomness, which means that it plays nicely with scan_emit's checkpointed rrule.

TemporalGPs.destructure — Method

destructure(x::AbstractVector, ys::AbstractVector)

Construct a flattened representation of ys in accordance with the structure in x.

TemporalGPs.dtcify — Method

dtcify(z::AbstractVector, x)

Recurse through x and replace any Separable kernels with a DTCSeparable. The resulting kernel can be used to construct the approximate model utilised by the DTC, and is used to compute the ELBO.

TemporalGPs.get_times — Method

get_times(x::AbstractVector{<:Real})

Get the times associated with the inputs.

TemporalGPs.inputs_to_time_form — Method

inputs_to_time_form(x::AbstractVector)

Get the time-form representation of a vector of inputs. Outputs an AbstractVector whose length is length(get_times(x)), and whose elements comprise all of those associated with each time point.

For single-output time series problems, x is usually returned as-is, however, for multi-output and spatio-temporal problems there are usually multiple inputs associated with each point in time.

TemporalGPs.marginals_diag — Method

marginals_diag(model::LGSSM)

Compute the diagonal of the marginals at each point in time. These are returned as a Vector of length length(model), each element of which is a diagonal Gaussian.

TemporalGPs.merge_inputs — Method

merge_inputs(x1::AbstractVector, x2::AbstractVector)

Merge two collections of inputs x1 and x2 if possible. A notable example of inputs which cannot be merged are two rectilinear grids whose spatial components differ.

TemporalGPs.noise_var_to_time_form — Method

noise_var_to_time_form(x::AbstractVector, S::AbstractMatrix{<:Real})

Get the time-form representation of an AbstractMatrix of observation noise variances S, based on the AbstractVector of inputs x.

For example, single-output time problems will typically require S to be Diagonal, and will just return its diagonal. Multi-output and spatio-temporal problems can potentially have a block-diagonal S, as it's typically only necessary to require that the observation variance at each point in time is independent of other times.

TemporalGPs.observations_to_time_form — Method

observations_to_time_form(x::AbstractVector, y::AbstractVector{<:Real})

Get the time-form representation of a vector of observations y of length length(time_points(x)), which is useful for state-space models. The precise arrangements of the elements of y in the output depends on x.

For example, single-output time-series problems typically just return y, but multi-output and spatio-temporal problems typically have multiple elements of y associated with a single element of x.

TemporalGPs.predict — Method

predict(x::Gaussian, f::AbstractLGC)::Gaussian{Tm,AbstractMatrix}

Compute the distribution "predicted" by this conditional given a Gaussian input x. Will be equivalent to

    Gaussian(f.A * x.m + f.a, f.A * x.P * f.A' + f.Q)

TemporalGPs.predict_marginals — Method

predict_marginals(x::Gaussian, f::AbstractLGC)::Gaussian{Tm,Diagonal}

Equivalent to

    xꜝ⁺¹ = predict(xꜝ, f)
    Gaussian(mean(xꜝ⁺¹), Diagonal(cov(xꜝ⁺¹)))

TemporalGPs.scan_emit — Method

Like Transducers.ScanEmit, but `f` isn't allowed to have internal state, and slightly
faster in some cases that I care about for this package.

This function makes some strong assumptions about the nature of `f` and `xs` so that it
can achieve good performance. In particular, `f` must not involve any globals, and
the output of `f` should not change type for different elements of `x`.

TemporalGPs.sort_in_time — Method

sort_in_time(x::AbstractVector)

Sort the input x temporally. This is different from sorting the input in general, but agrees for real-valued inputs. Returns both the sorting indices and the sorted inputs.

TemporalGPs.to_sde — Function

to_sde(f::GP, storage::StorageType)

Return a LTISDE wrapper around f that allows TemporalGPs to perform state-space inference.

TemporalGPs.to_sde — Method

to_sde(k::Kernel, storage_type::StorageType)

Return the equivalent state-space representation of the kernel k as a tuple (F, q, H), where F is the state transition matrix, q is the process noise and H is the observation matrix.

TemporalGPs.zygote_friendly_map — Method

zygote_friendly_map(f, x)

This version of map is a bit weird. It makes slightly stronger assumptions about the nature of what you're allowed to pass in to it than Base.map does and, in return, you get much improved performance when used in conjunction with Zygote.

Assumptions.

No globals are used in f. This means that TemporalGPs.NoContext can be employed.

f has no fields. If you've got data to share across elements, use a Fill.

Similarly, f has no mutable state (follows from the above).

f doesn't mutate its argument.

TemporalGPs.ε_randn — Method

ε_randn(rng::AbstractRNG, f::AbstractLGC)

Generate the vector of random numbers needed inside conditional_rand.